fix: moved AcceptedGiftTypes.kt to the appropriate package
